Fix comparison of signed/unsigned ints in RootGeoManager
Problem is std::list uses unsigned, and node->GetNDaughters() returns
signed ints. In the future, it will be necessary to decide the return
type of size() member functions and adopt a common standard everywhere,
to avoid warnings about comparison of signed and unsigned integers.

+++ source/RootGeoManager.cpp
@@ -130,7 +130,7 @@ VPlacedVolume* RootGeoManager::Convert(TGeoNode const *const node) {
     remaining_daughters = node->GetNdaughters()
                           - logical_volume->GetDaughters().size();
     assert(remaining_daughters == 0 ||
-           remaining_daughters == node->GetNdaughters());
+           remaining_daughters == (unsigned int) node->GetNdaughters());
   }
 
   // we have to convert here assemblies to list of normal nodes
